split today question section
