+++ frontend/src/Components/TodayQuestion/TodayQuestion.elements.js
@@ -1,0 +1,64 @@
+import styled from "styled-components";
+
+export const TodayLeftContainer = styled.div`
+  width: 50%;
+`;
+
+export const TodayRightContainer = styled.section`
+  width: 50%;
+  margin-left: 1%;
+  display: flex;
+  flex-direction: column;
+  justify-content: space-between;
+`;
+
+export const QuestionArea = styled.section`
+  font-size: 3rem;
+`;
+
+export const ShuffleButton = styled.div`
+  cursor: pointer;
+`;
+
+export const TodayQuestion = styled.div`
+  margin-top: 5%;
+`;
+
+export const QDate = styled.h5`
+  font-size: 1.2rem;
+`;
+
+export const Question = styled.h2`
+  font-size: 3rem;
+`;
+
+export const PostCardSArea = styled.section`
+  display: flex;
+  flex-direction: column;
+  text-align: center;
+`;
+
+export const TodaySTop = styled.div`
+  display: flex;
+  align-items: center;
+  gap: 5%;
+`;
+
+export const PostCardSWrapper = styled.section`
+  width: 80%;
+  margin: 5% 0;
+  display: flex;
+  justify-content: center;
+  gap: 1%;
+`;
+
+export const IndicatorWrapper = styled.div`
+  display: flex;
+  justify-content: center;
+  gap: 5%;
+`;
+
+export const LoadMoreButtonContainer = styled.div`
+  width: 5%;
+  cursor: pointer;
+`;
