+++ frontend/src/Components/index.js
@@ -10,3 +10,4 @@ export { default as CreateCard } from "./CreateCard/CreateCard";
 export { default as QuestionCard } from "./QuestionCard/QuestionCard";
 export { default as QuestionCardGrid } from "./QuestionCardGrid/QuestionCardGrid";
 export { default as QuestionOpened } from "./QuestionOpened/QuestionOpened";
+export { default as TodayQuestion } from "./TodayQuestion/TodayQuestion";
