+++ frontend/static/frontend/main.js
@@ -668,6 +668,28 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 
 /***/ }),
 
+/***/ "./src/Components/TodayQuestion/TodayQuestion.elements.js":
+/*!****************************************************************!*\
+  !*** ./src/Components/TodayQuestion/TodayQuestion.elements.js ***!
+  \****************************************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"TodayLeftContainer\": () => (/* binding */ TodayLeftContainer),\n/* harmony export */   \"TodayRightContainer\": () => (/* binding */ TodayRightContainer),\n/* harmony export */   \"QuestionArea\": () => (/* binding */ QuestionArea),\n/* harmony export */   \"ShuffleButton\": () => (/* binding */ ShuffleButton),\n/* harmony export */   \"TodayQuestion\": () => (/* binding */ TodayQuestion),\n/* harmony export */   \"QDate\": () => (/* binding */ QDate),\n/* harmony export */   \"Question\": () => (/* binding */ Question),\n/* harmony export */   \"PostCardSArea\": () => (/* binding */ PostCardSArea),\n/* harmony export */   \"TodaySTop\": () => (/* binding */ TodaySTop),\n/* harmony export */   \"PostCardSWrapper\": () => (/* binding */ PostCardSWrapper),\n/* harmony export */   \"IndicatorWrapper\": () => (/* binding */ IndicatorWrapper),\n/* harmony export */   \"LoadMoreButtonContainer\": () => (/* binding */ LoadMoreButtonContainer)\n/* harmony export */ });\n/* harmony import */ var styled_components__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! styled-components */ \"./node_modules/styled-components/dist/styled-components.browser.esm.js\");\nvar _templateObject, _templateObject2, _templateObject3, _templateObject4, _templateObject5, _templateObject6, _templateObject7, _templateObject8, _templateObject9, _templateObject10, _templateObject11, _templateObject12;\n\nfunction _taggedTemplateLiteral(strings, raw) { if (!raw) { raw = strings.slice(0); } return Object.freeze(Object.defineProperties(strings, { raw: { value: Object.freeze(raw) } })); }\n\n\nvar TodayLeftContainer = styled_components__WEBPACK_IMPORTED_MODULE_0__[\"default\"].div(_templateObject || (_templateObject = _taggedTemplateLiteral([\"\\n  width: 50%;\\n\"])));\nvar TodayRightContainer = styled_components__WEBPACK_IMPORTED_MODULE_0__[\"default\"].section(_templateObject2 || (_templateObject2 = _taggedTemplateLiteral([\"\\n  width: 50%;\\n  margin-left: 1%;\\n  display: flex;\\n  flex-direction: column;\\n  justify-content: space-between;\\n\"])));\nvar QuestionArea = styled_components__WEBPACK_IMPORTED_MODULE_0__[\"default\"].section(_templateObject3 || (_templateObject3 = _taggedTemplateLiteral([\"\\n  font-size: 3rem;\\n\"])));\nvar ShuffleButton = styled_components__WEBPACK_IMPORTED_MODULE_0__[\"default\"].div(_templateObject4 || (_templateObject4 = _taggedTemplateLiteral([\"\\n  cursor: pointer;\\n\"])));\nvar TodayQuestion = styled_components__WEBPACK_IMPORTED_MODULE_0__[\"default\"].div(_templateObject5 || (_templateObject5 = _taggedTemplateLiteral([\"\\n  margin-top: 5%;\\n\"])));\nvar QDate = styled_components__WEBPACK_IMPORTED_MODULE_0__[\"default\"].h5(_templateObject6 || (_templateObject6 = _taggedTemplateLiteral([\"\\n  font-size: 1.2rem;\\n\"])));\nvar Question = styled_components__WEBPACK_IMPORTED_MODULE_0__[\"default\"].h2(_templateObject7 || (_templateObject7 = _taggedTemplateLiteral([\"\\n  font-size: 3rem;\\n\"])));\nvar PostCardSArea = styled_components__WEBPACK_IMPORTED_MODULE_0__[\"default\"].section(_templateObject8 || (_templateObject8 = _taggedTemplateLiteral([\"\\n  display: flex;\\n  flex-direction: column;\\n  text-align: center;\\n\"])));\nvar TodaySTop = styled_components__WEBPACK_IMPORTED_MODULE_0__[\"default\"].div(_templateObject9 || (_templateObject9 = _taggedTemplateLiteral([\"\\n  display: flex;\\n  align-items: center;\\n  gap: 5%;\\n\"])));\nvar PostCardSWrapper = styled_components__WEBPACK_IMPORTED_MODULE_0__[\"default\"].section(_templateObject10 || (_templateObject10 = _taggedTemplateLiteral([\"\\n  width: 80%;\\n  margin: 5% 0;\\n  display: flex;\\n  justify-content: center;\\n  gap: 1%;\\n\"])));\nvar IndicatorWrapper = styled_components__WEBPACK_IMPORTED_MODULE_0__[\"default\"].div(_templateObject11 || (_templateObject11 = _taggedTemplateLiteral([\"\\n  display: flex;\\n  justify-content: center;\\n  gap: 5%;\\n\"])));\nvar LoadMoreButtonContainer = styled_components__WEBPACK_IMPORTED_MODULE_0__[\"default\"].div(_templateObject12 || (_templateObject12 = _taggedTemplateLiteral([\"\\n  width: 5%;\\n  cursor: pointer;\\n\"])));\n\n//# sourceURL=webpack://frontend/./src/Components/TodayQuestion/TodayQuestion.elements.js?");
+
+/***/ }),
+
+/***/ "./src/Components/TodayQuestion/TodayQuestion.js":
+/*!*******************************************************!*\
+  !*** ./src/Components/TodayQuestion/TodayQuestion.js ***!
+  \*******************************************************/
+/***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
+
+"use strict";
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var _TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./TodayQuestion.elements */ \"./src/Components/TodayQuestion/TodayQuestion.elements.js\");\n/* harmony import */ var _globalStyles__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../globalStyles */ \"./src/globalStyles.js\");\n/* harmony import */ var _Components__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../../Components */ \"./src/Components/index.js\");\nfunction _createForOfIteratorHelper(o, allowArrayLike) { var it = typeof Symbol !== \"undefined\" && o[Symbol.iterator] || o[\"@@iterator\"]; if (!it) { if (Array.isArray(o) || (it = _unsupportedIterableToArray(o)) || allowArrayLike && o && typeof o.length === \"number\") { if (it) o = it; var i = 0; var F = function F() {}; return { s: F, n: function n() { if (i >= o.length) return { done: true }; return { done: false, value: o[i++] }; }, e: function e(_e2) { throw _e2; }, f: F }; } throw new TypeError(\"Invalid attempt to iterate non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); } var normalCompletion = true, didErr = false, err; return { s: function s() { it = it.call(o); }, n: function n() { var step = it.next(); normalCompletion = step.done; return step; }, e: function e(_e3) { didErr = true; err = _e3; }, f: function f() { try { if (!normalCompletion && it[\"return\"] != null) it[\"return\"](); } finally { if (didErr) throw err; } } }; }\n\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\n\nvar TodayQuestion = function TodayQuestion(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId;\n\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState2 = _slicedToArray(_useState, 2),\n      content = _useState2[0],\n      setContent = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState4 = _slicedToArray(_useState3, 2),\n      questionList = _useState4[0],\n      setQList = _useState4[1];\n\n  var _useState5 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(),\n      _useState6 = _slicedToArray(_useState5, 2),\n      question = _useState6[0],\n      setQuestion = _useState6[1];\n\n  var _useState7 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(),\n      _useState8 = _slicedToArray(_useState7, 2),\n      questionId = _useState8[0],\n      setQuestionId = _useState8[1];\n\n  var _useState9 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState10 = _slicedToArray(_useState9, 2),\n      todayQCards = _useState10[0],\n      setTodayQCards = _useState10[1];\n\n  var _useState11 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState12 = _slicedToArray(_useState11, 2),\n      indicators = _useState12[0],\n      setIndicators = _useState12[1];\n\n  var _useState13 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(),\n      _useState14 = _slicedToArray(_useState13, 2),\n      contentIdx = _useState14[0],\n      setIdx = _useState14[1];\n\n  var _useState15 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(),\n      _useState16 = _slicedToArray(_useState15, 2),\n      cardLId = _useState16[0],\n      setCardLId = _useState16[1]; // on Mount\n\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    axios__WEBPACK_IMPORTED_MODULE_1___default()({\n      method: \"get\",\n      url: \"/api/question/past\",\n      headers: {\n        \"Content-Type\": \"application/json\"\n      }\n    }).then(function (response) {\n      if (response.status > 400) {\n        setPlaceholder(\"Something went wrong!\");\n      }\n\n      return response.data;\n    }).then(function (data) {\n      setQList(data);\n    });\n  }, []); // on question list change\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    if (questionList[0]) {\n      setQuestion(questionList[0].question_content);\n      setQuestionId(questionList[0].id);\n    }\n  }, [questionList]); // get 10 post of selected question\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    if (questionId) {\n      axios__WEBPACK_IMPORTED_MODULE_1___default()({\n        method: \"get\",\n        url: \"api/postlist-question/\".concat(questionId, \"/10\"),\n        headers: {\n          \"Content-Type\": \"application/json\"\n        }\n      }).then(function (response) {\n        if (response.status > 400) {\n          setPlaceholder(\"Something went wrong!\");\n        }\n\n        return response.data;\n      }).then(function (data) {\n        setContent(data);\n      });\n    }\n  }, [questionId]); // when content for today question is updated\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    var arr = [];\n\n    var _iterator = _createForOfIteratorHelper(content),\n        _step;\n\n    try {\n      for (_iterator.s(); !(_step = _iterator.n()).done;) {\n        var c = _step.value;\n        arr.push( /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.PostCardS, {\n          key: c.id,\n          id: c.id,\n          user: c.user,\n          title: c.title,\n          track_title: c.track_title,\n          track_artist: c.track_artist,\n          track_album_cover: c.track_album_cover,\n          handleCardOpen: switchCardL\n        }));\n      }\n    } catch (err) {\n      _iterator.e(err);\n    } finally {\n      _iterator.f();\n    }\n\n    setTodayQCards([].concat(arr));\n    setIdx(0);\n\n    if (content[0]) {\n      setCardLId(content[0].id);\n    }\n  }, [content]); // when content is updated or idx of CardS has changed\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    var indArr = [];\n\n    for (var i = 0; i < Math.ceil(content.length / 2); i++) {\n      indArr.push( /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_globalStyles__WEBPACK_IMPORTED_MODULE_3__.IndicatorDot, {\n        key: i,\n        selected: i * 2 === contentIdx\n      }));\n    }\n\n    setIndicators([].concat(indArr));\n  }, [contentIdx, content]); // update CardS idx\n\n  var handleTodayQCard = function handleTodayQCard(d) {\n    var idx;\n\n    if (d > 0) {\n      idx = contentIdx === todayQCards.length - 1 || contentIdx + 1 === todayQCards.length - 1 ? 0 : contentIdx + 2;\n    } else {\n      idx = contentIdx === 0 ? todayQCards.length - 1 : contentIdx - 2;\n    }\n\n    setIdx(idx);\n  }; // generate random idx for question\n\n\n  var randomQuestion = function randomQuestion() {\n    setContent([]);\n    setTodayQCards([]);\n    var qIdx;\n\n    do {\n      qIdx = Math.floor(Math.random() * 100) % questionList.length;\n    } while (questionList[qIdx].id === questionId);\n\n    setQuestion(questionList[qIdx].question_content);\n    setQuestionId(questionList[qIdx].id);\n  }; // CardL switch when CardS clicked\n\n\n  var switchCardL = function switchCardL(id) {\n    setCardLId(id);\n  };\n\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.TodayLeftContainer, null, content.length ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.PostCardL, {\n    postId: cardLId,\n    currUser: currUser,\n    userId: userId,\n    token: token\n  }) : \"\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.TodayRightContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.QuestionArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.ShuffleButton, {\n    onClick: randomQuestion\n  }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"i\", {\n    className: \"fas fa-random\"\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.TodayQuestion, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.QDate, null, \"\\uC624\\uB298\\uC758 \\uC9C8\\uBB38\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.Question, null, question))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.PostCardSArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.TodaySTop, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.LoadMoreButtonContainer, {\n    onClick: function onClick() {\n      return handleTodayQCard(-1);\n    }\n  }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"i\", {\n    className: \"fas fa-chevron-left\"\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.PostCardSWrapper, null, todayQCards[contentIdx], contentIdx + 1 < todayQCards.length - 1 ? todayQCards[contentIdx + 1] : \"\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.LoadMoreButtonContainer, {\n    onClick: function onClick() {\n      return handleTodayQCard(1);\n    }\n  }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"i\", {\n    className: \"fas fa-chevron-right\"\n  }))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_TodayQuestion_elements__WEBPACK_IMPORTED_MODULE_2__.IndicatorWrapper, null, indicators))));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (TodayQuestion);\n\n//# sourceURL=webpack://frontend/./src/Components/TodayQuestion/TodayQuestion.js?");
+
+/***/ }),
+
 /***/ "./src/Components/csrftoken.js":
 /*!*************************************!*\
   !*** ./src/Components/csrftoken.js ***!
@@ -686,7 +708,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"MusicPlayer\": () => (/* reexport safe */ _MusicPlayer_MusicPlayer__WEBPACK_IMPORTED_MODULE_0__[\"default\"]),\n/* harmony export */   \"MusicCard\": () => (/* reexport safe */ _MusicCard_MusicCard__WEBPACK_IMPORTED_MODULE_1__[\"default\"]),\n/* harmony export */   \"Calendar\": () => (/* reexport safe */ _Calendar_Calendar__WEBPACK_IMPORTED_MODULE_2__[\"default\"]),\n/* harmony export */   \"List\": () => (/* reexport safe */ _List_List__WEBPACK_IMPORTED_MODULE_3__[\"default\"]),\n/* harmony export */   \"DiaryForm\": () => (/* reexport safe */ _DiaryForm_DiaryForm__WEBPACK_IMPORTED_MODULE_4__[\"default\"]),\n/* harmony export */   \"SearchedItem\": () => (/* reexport safe */ _SearchedItem_SearchedItem__WEBPACK_IMPORTED_MODULE_5__[\"default\"]),\n/* harmony export */   \"PostCardL\": () => (/* reexport safe */ _PostCardL_PostCardL__WEBPACK_IMPORTED_MODULE_6__[\"default\"]),\n/* harmony export */   \"PostCardS\": () => (/* reexport safe */ _PostCardS_PostCardS__WEBPACK_IMPORTED_MODULE_7__[\"default\"]),\n/* harmony export */   \"CreateCard\": () => (/* reexport safe */ _CreateCard_CreateCard__WEBPACK_IMPORTED_MODULE_8__[\"default\"]),\n/* harmony export */   \"QuestionCard\": () => (/* reexport safe */ _QuestionCard_QuestionCard__WEBPACK_IMPORTED_MODULE_9__[\"default\"]),\n/* harmony export */   \"QuestionCardGrid\": () => (/* reexport safe */ _QuestionCardGrid_QuestionCardGrid__WEBPACK_IMPORTED_MODULE_10__[\"default\"]),\n/* harmony export */   \"QuestionOpened\": () => (/* reexport safe */ _QuestionOpened_QuestionOpened__WEBPACK_IMPORTED_MODULE_11__[\"default\"])\n/* harmony export */ });\n/* harmony import */ var _MusicPlayer_MusicPlayer__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./MusicPlayer/MusicPlayer */ \"./src/Components/MusicPlayer/MusicPlayer.js\");\n/* harmony import */ var _MusicCard_MusicCard__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./MusicCard/MusicCard */ \"./src/Components/MusicCard/MusicCard.js\");\n/* harmony import */ var _Calendar_Calendar__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./Calendar/Calendar */ \"./src/Components/Calendar/Calendar.js\");\n/* harmony import */ var _List_List__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./List/List */ \"./src/Components/List/List.js\");\n/* harmony import */ var _DiaryForm_DiaryForm__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./DiaryForm/DiaryForm */ \"./src/Components/DiaryForm/DiaryForm.js\");\n/* harmony import */ var _SearchedItem_SearchedItem__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./SearchedItem/SearchedItem */ \"./src/Components/SearchedItem/SearchedItem.js\");\n/* harmony import */ var _PostCardL_PostCardL__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./PostCardL/PostCardL */ \"./src/Components/PostCardL/PostCardL.js\");\n/* harmony import */ var _PostCardS_PostCardS__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./PostCardS/PostCardS */ \"./src/Components/PostCardS/PostCardS.js\");\n/* harmony import */ var _CreateCard_CreateCard__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! ./CreateCard/CreateCard */ \"./src/Components/CreateCard/CreateCard.js\");\n/* harmony import */ var _QuestionCard_QuestionCard__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! ./QuestionCard/QuestionCard */ \"./src/Components/QuestionCard/QuestionCard.js\");\n/* harmony import */ var _QuestionCardGrid_QuestionCardGrid__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! ./QuestionCardGrid/QuestionCardGrid */ \"./src/Components/QuestionCardGrid/QuestionCardGrid.js\");\n/* harmony import */ var _QuestionOpened_QuestionOpened__WEBPACK_IMPORTED_MODULE_11__ = __webpack_require__(/*! ./QuestionOpened/QuestionOpened */ \"./src/Components/QuestionOpened/QuestionOpened.js\");\n\n\n\n\n\n\n\n\n\n\n\n\n\n//# sourceURL=webpack://frontend/./src/Components/index.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"MusicPlayer\": () => (/* reexport safe */ _MusicPlayer_MusicPlayer__WEBPACK_IMPORTED_MODULE_0__[\"default\"]),\n/* harmony export */   \"MusicCard\": () => (/* reexport safe */ _MusicCard_MusicCard__WEBPACK_IMPORTED_MODULE_1__[\"default\"]),\n/* harmony export */   \"Calendar\": () => (/* reexport safe */ _Calendar_Calendar__WEBPACK_IMPORTED_MODULE_2__[\"default\"]),\n/* harmony export */   \"List\": () => (/* reexport safe */ _List_List__WEBPACK_IMPORTED_MODULE_3__[\"default\"]),\n/* harmony export */   \"DiaryForm\": () => (/* reexport safe */ _DiaryForm_DiaryForm__WEBPACK_IMPORTED_MODULE_4__[\"default\"]),\n/* harmony export */   \"SearchedItem\": () => (/* reexport safe */ _SearchedItem_SearchedItem__WEBPACK_IMPORTED_MODULE_5__[\"default\"]),\n/* harmony export */   \"PostCardL\": () => (/* reexport safe */ _PostCardL_PostCardL__WEBPACK_IMPORTED_MODULE_6__[\"default\"]),\n/* harmony export */   \"PostCardS\": () => (/* reexport safe */ _PostCardS_PostCardS__WEBPACK_IMPORTED_MODULE_7__[\"default\"]),\n/* harmony export */   \"CreateCard\": () => (/* reexport safe */ _CreateCard_CreateCard__WEBPACK_IMPORTED_MODULE_8__[\"default\"]),\n/* harmony export */   \"QuestionCard\": () => (/* reexport safe */ _QuestionCard_QuestionCard__WEBPACK_IMPORTED_MODULE_9__[\"default\"]),\n/* harmony export */   \"QuestionCardGrid\": () => (/* reexport safe */ _QuestionCardGrid_QuestionCardGrid__WEBPACK_IMPORTED_MODULE_10__[\"default\"]),\n/* harmony export */   \"QuestionOpened\": () => (/* reexport safe */ _QuestionOpened_QuestionOpened__WEBPACK_IMPORTED_MODULE_11__[\"default\"]),\n/* harmony export */   \"TodayQuestion\": () => (/* reexport safe */ _TodayQuestion_TodayQuestion__WEBPACK_IMPORTED_MODULE_12__[\"default\"])\n/* harmony export */ });\n/* harmony import */ var _MusicPlayer_MusicPlayer__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! ./MusicPlayer/MusicPlayer */ \"./src/Components/MusicPlayer/MusicPlayer.js\");\n/* harmony import */ var _MusicCard_MusicCard__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! ./MusicCard/MusicCard */ \"./src/Components/MusicCard/MusicCard.js\");\n/* harmony import */ var _Calendar_Calendar__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./Calendar/Calendar */ \"./src/Components/Calendar/Calendar.js\");\n/* harmony import */ var _List_List__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ./List/List */ \"./src/Components/List/List.js\");\n/* harmony import */ var _DiaryForm_DiaryForm__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ./DiaryForm/DiaryForm */ \"./src/Components/DiaryForm/DiaryForm.js\");\n/* harmony import */ var _SearchedItem_SearchedItem__WEBPACK_IMPORTED_MODULE_5__ = __webpack_require__(/*! ./SearchedItem/SearchedItem */ \"./src/Components/SearchedItem/SearchedItem.js\");\n/* harmony import */ var _PostCardL_PostCardL__WEBPACK_IMPORTED_MODULE_6__ = __webpack_require__(/*! ./PostCardL/PostCardL */ \"./src/Components/PostCardL/PostCardL.js\");\n/* harmony import */ var _PostCardS_PostCardS__WEBPACK_IMPORTED_MODULE_7__ = __webpack_require__(/*! ./PostCardS/PostCardS */ \"./src/Components/PostCardS/PostCardS.js\");\n/* harmony import */ var _CreateCard_CreateCard__WEBPACK_IMPORTED_MODULE_8__ = __webpack_require__(/*! ./CreateCard/CreateCard */ \"./src/Components/CreateCard/CreateCard.js\");\n/* harmony import */ var _QuestionCard_QuestionCard__WEBPACK_IMPORTED_MODULE_9__ = __webpack_require__(/*! ./QuestionCard/QuestionCard */ \"./src/Components/QuestionCard/QuestionCard.js\");\n/* harmony import */ var _QuestionCardGrid_QuestionCardGrid__WEBPACK_IMPORTED_MODULE_10__ = __webpack_require__(/*! ./QuestionCardGrid/QuestionCardGrid */ \"./src/Components/QuestionCardGrid/QuestionCardGrid.js\");\n/* harmony import */ var _QuestionOpened_QuestionOpened__WEBPACK_IMPORTED_MODULE_11__ = __webpack_require__(/*! ./QuestionOpened/QuestionOpened */ \"./src/Components/QuestionOpened/QuestionOpened.js\");\n/* harmony import */ var _TodayQuestion_TodayQuestion__WEBPACK_IMPORTED_MODULE_12__ = __webpack_require__(/*! ./TodayQuestion/TodayQuestion */ \"./src/Components/TodayQuestion/TodayQuestion.js\");\n\n\n\n\n\n\n\n\n\n\n\n\n\n\n//# sourceURL=webpack://frontend/./src/Components/index.js?");
 
 /***/ }),
 
@@ -774,7 +796,7 @@ eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpac
 /***/ ((__unused_webpack_module, __webpack_exports__, __webpack_require__) => {
 
 "use strict";
-eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var _Home_elements__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./Home.elements */ \"./src/Containers/Home/Home.elements.js\");\n/* harmony import */ var _Components__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../Components */ \"./src/Components/index.js\");\n/* harmony import */ var _globalStyles__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../../globalStyles */ \"./src/globalStyles.js\");\nfunction _createForOfIteratorHelper(o, allowArrayLike) { var it = typeof Symbol !== \"undefined\" && o[Symbol.iterator] || o[\"@@iterator\"]; if (!it) { if (Array.isArray(o) || (it = _unsupportedIterableToArray(o)) || allowArrayLike && o && typeof o.length === \"number\") { if (it) o = it; var i = 0; var F = function F() {}; return { s: F, n: function n() { if (i >= o.length) return { done: true }; return { done: false, value: o[i++] }; }, e: function e(_e2) { throw _e2; }, f: F }; } throw new TypeError(\"Invalid attempt to iterate non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); } var normalCompletion = true, didErr = false, err; return { s: function s() { it = it.call(o); }, n: function n() { var step = it.next(); normalCompletion = step.done; return step; }, e: function e(_e3) { didErr = true; err = _e3; }, f: function f() { try { if (!normalCompletion && it[\"return\"] != null) it[\"return\"](); } finally { if (didErr) throw err; } } }; }\n\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\n\n\nvar Home = function Home(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId,\n      postId = _ref.postId;\n\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState2 = _slicedToArray(_useState, 2),\n      content = _useState2[0],\n      setContent = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState4 = _slicedToArray(_useState3, 2),\n      questionList = _useState4[0],\n      setQList = _useState4[1];\n\n  var _useState5 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(),\n      _useState6 = _slicedToArray(_useState5, 2),\n      question = _useState6[0],\n      setQuestion = _useState6[1];\n\n  var _useState7 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(),\n      _useState8 = _slicedToArray(_useState7, 2),\n      questionId = _useState8[0],\n      setQuestionId = _useState8[1];\n\n  var _useState9 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState10 = _slicedToArray(_useState9, 2),\n      todayQCards = _useState10[0],\n      setTodayQCards = _useState10[1];\n\n  var _useState11 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)([]),\n      _useState12 = _slicedToArray(_useState11, 2),\n      indicators = _useState12[0],\n      setIndicators = _useState12[1];\n\n  var _useState13 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(),\n      _useState14 = _slicedToArray(_useState13, 2),\n      contentIdx = _useState14[0],\n      setIdx = _useState14[1];\n\n  var _useState15 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(),\n      _useState16 = _slicedToArray(_useState15, 2),\n      cardLId = _useState16[0],\n      setCardLId = _useState16[1];\n\n  var _useState17 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(false),\n      _useState18 = _slicedToArray(_useState17, 2),\n      loaded = _useState18[0],\n      setLoad = _useState18[1];\n\n  var _useState19 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(\"Loading Content\"),\n      _useState20 = _slicedToArray(_useState19, 2),\n      placeholder = _useState20[0],\n      setPlaceholder = _useState20[1]; // on Mount\n\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    axios__WEBPACK_IMPORTED_MODULE_1___default()({\n      method: \"get\",\n      url: \"/api/question/past\",\n      headers: {\n        \"Content-Type\": \"application/json\"\n      }\n    }).then(function (response) {\n      if (response.status > 400) {\n        setPlaceholder(\"Something went wrong!\");\n      }\n\n      return response.data;\n    }).then(function (data) {\n      setQList(data);\n    });\n  }, []); // on question list change\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    if (questionList[0]) {\n      setQuestion(questionList[0].question_content);\n      setQuestionId(questionList[0].id);\n    }\n  }, [questionList]); // get 10 post of selected question\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    if (questionId) {\n      axios__WEBPACK_IMPORTED_MODULE_1___default()({\n        method: \"get\",\n        url: \"api/postlist-question/\".concat(questionId, \"/10\"),\n        headers: {\n          \"Content-Type\": \"application/json\"\n        }\n      }).then(function (response) {\n        if (response.status > 400) {\n          setPlaceholder(\"Something went wrong!\");\n        }\n\n        return response.data;\n      }).then(function (data) {\n        setContent(data);\n      }).then(function () {\n        setLoad(true);\n      });\n    }\n  }, [questionId]); // when content for today question is updated\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    var arr = [];\n\n    var _iterator = _createForOfIteratorHelper(content),\n        _step;\n\n    try {\n      for (_iterator.s(); !(_step = _iterator.n()).done;) {\n        var c = _step.value;\n        arr.push( /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_3__.PostCardS, {\n          key: c.id,\n          id: c.id,\n          user: c.user,\n          title: c.title,\n          track_title: c.track_title,\n          track_artist: c.track_artist,\n          track_album_cover: c.track_album_cover,\n          handleCardOpen: switchCardL\n        }));\n      }\n    } catch (err) {\n      _iterator.e(err);\n    } finally {\n      _iterator.f();\n    }\n\n    setTodayQCards([].concat(arr));\n    setIdx(0);\n\n    if (content[0]) {\n      setCardLId(content[0].id);\n    }\n  }, [content]); // when content is updated or idx of CardS has changed\n\n  (0,react__WEBPACK_IMPORTED_MODULE_0__.useEffect)(function () {\n    var indArr = [];\n\n    for (var i = 0; i < Math.ceil(content.length / 2); i++) {\n      indArr.push( /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_globalStyles__WEBPACK_IMPORTED_MODULE_4__.IndicatorDot, {\n        key: i,\n        selected: i * 2 === contentIdx\n      }));\n    }\n\n    setIndicators([].concat(indArr));\n  }, [contentIdx, content]); // update CardS idx\n\n  var handleTodayQCard = function handleTodayQCard(d) {\n    var idx;\n\n    if (d > 0) {\n      idx = contentIdx === todayQCards.length - 1 || contentIdx + 1 === todayQCards.length - 1 ? 0 : contentIdx + 2;\n    } else {\n      idx = contentIdx === 0 ? todayQCards.length - 1 : contentIdx - 2;\n    }\n\n    setIdx(idx);\n  }; // generate random idx for question\n\n\n  var randomQuestion = function randomQuestion() {\n    setContent([]);\n    setTodayQCards([]);\n    var qIdx;\n\n    do {\n      qIdx = Math.floor(Math.random() * 100) % questionList.length;\n    } while (questionList[qIdx].id === questionId);\n\n    setQuestion(questionList[qIdx].question_content);\n    setQuestionId(questionList[qIdx].id);\n  }; // CardL switch when CardS clicked\n\n\n  var switchCardL = function switchCardL(id) {\n    setCardLId(id);\n  };\n\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.TodayQuestionSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.TodayLeftContainer, null, content.length ? /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_3__.PostCardL, {\n    postId: cardLId,\n    currUser: currUser,\n    userId: userId,\n    token: token\n  }) : \"\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.TodayRightContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.QuestionArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.ShuffleButton, {\n    onClick: randomQuestion\n  }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"i\", {\n    className: \"fas fa-random\"\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.TodayQuestion, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.QDate, null, \"\\uC624\\uB298\\uC758 \\uC9C8\\uBB38\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.Question, null, question))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PostCardSArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.TodaySTop, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.LoadMoreButtonContainer, {\n    onClick: function onClick() {\n      return handleTodayQCard(-1);\n    }\n  }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"i\", {\n    className: \"fas fa-chevron-left\"\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PostCardSWrapper, null, todayQCards[contentIdx], contentIdx + 1 < todayQCards.length - 1 ? todayQCards[contentIdx + 1] : \"\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.LoadMoreButtonContainer, {\n    onClick: function onClick() {\n      return handleTodayQCard(1);\n    }\n  }, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"i\", {\n    className: \"fas fa-chevron-right\"\n  }))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.IndicatorWrapper, null, indicators)))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.CreatePageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.CreateCardLeft, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperLeft, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB2F9\\uC2E0\\uC758 \\uC774\\uC57C\\uAE30\\uB97C \\uB4E4\\uB824\\uC8FC\\uC138\\uC694. \\uC5B4\\uB5A4 \\uB0B4\\uC6A9\\uC774\\uB4E0 \\uC88B\\uC544\\uC694. \\uC544\\uC8FC \\uC0AC\\uC18C\\uD55C \\uAC83\\uBD80\\uD130 \\uAE4A\\uC740 \\uC18D\\uB9C8\\uC74C\\uAE4C\\uC9C0, \\uB5A0\\uC624\\uB974\\uB294 \\uB300\\uB85C \\uC801\\uC5B4\\uBCFC\\uAE4C\\uC694?\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_3__.CreateCard, null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.CreateCardRight, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperQuestionArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperRight, null, \"\\uC9C8\\uBB38\\uC744 \\uB4E3\\uACE0,\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB5A0\\uC624\\uB974\\uB294 \\uC74C\\uC545\\uC774\\uB098 \\uC774\\uC57C\\uAE30\\uAC00 \\uC788\\uB098\\uC694?\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB2F9\\uC2E0\\uC758 \\uC774\\uC57C\\uAE30\\uC5D0 \\uC5B4\\uC6B8\\uB9AC\\uB294 \\uC74C\\uC545\\uC744\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uC9C1\\uC811 \\uACE8\\uB77C\\uC8FC\\uC138\\uC694.\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperSearchArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperRight, null)), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperDoneArea, null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.DoneButtonArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperRight, null, \"\\uC774\\uC81C \\uB2F9\\uC2E0\\uC758 \\uC774\\uC57C\\uAE30\\uB97C\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB2E4\\uB978 \\uC0AC\\uB78C\\uB4E4\\uACFC \\uB098\\uB220\\uBCF4\\uC138\\uC694!\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null)), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_globalStyles__WEBPACK_IMPORTED_MODULE_4__.ButtonWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_globalStyles__WEBPACK_IMPORTED_MODULE_4__.DefaultButton, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_globalStyles__WEBPACK_IMPORTED_MODULE_4__.ButtonIconArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_globalStyles__WEBPACK_IMPORTED_MODULE_4__.ButtonIcon, {\n    className: \"fa fa-check\"\n  })), \"\\uB2E4 \\uC37C\\uC5B4\\uC694\"))))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.QuestionListPageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeaderContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeader, null, \"\\uB2E4\\uB978 \\uC9C8\\uBB38 \\uB458\\uB7EC\\uBCF4\\uAE30\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.ToQuestionListButton, null, \"\\uBAA8\\uB4E0 \\uC9C8\\uBB38 \\uBCF4\\uAE30 >\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_3__.QuestionCardGrid, {\n    currUser: currUser,\n    token: token,\n    userId: userId\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicListPageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeaderContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeader, null, \"\\uCD5C\\uADFC\\uC5D0 \\uC120\\uD0DD\\uB41C \\uC74C\\uC545\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicListContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicCardWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_3__.MusicCard, {\n    title: \"\\uC548\\uB155\",\n    artist: \"\\uC548\\uB155\",\n    source: \"\\uC548\\uB155\",\n    link: \"\\uC548\\uB155\",\n    cover: \"\\uC548\\uB155\",\n    cols: \"1\",\n    postId: postId\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicCardWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_3__.MusicCard, {\n    title: \"\\uC548\\uB155\",\n    artist: \"\\uC548\\uB155\",\n    source: \"\\uC548\\uB155\",\n    link: \"\\uC548\\uB155\",\n    cover: \"\\uC548\\uB155\",\n    cols: \"1\",\n    postId: postId\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicCardWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_3__.MusicCard, {\n    title: \"\\uC548\\uB155\",\n    artist: \"\\uC548\\uB155\",\n    source: \"\\uC548\\uB155\",\n    link: \"\\uC548\\uB155\",\n    cover: \"\\uC548\\uB155\",\n    cols: \"1\",\n    postId: postId\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicCardWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_3__.MusicCard, {\n    title: \"\\uC548\\uB155\",\n    artist: \"\\uC548\\uB155\",\n    source: \"\\uC548\\uB155\",\n    link: \"\\uC548\\uB155\",\n    cover: \"\\uC548\\uB155\",\n    cols: \"1\",\n    postId: postId\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicCardWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_3__.MusicCard, {\n    title: \"\\uC548\\uB155\",\n    artist: \"\\uC548\\uB155\",\n    source: \"\\uC548\\uB155\",\n    link: \"\\uC548\\uB155\",\n    cover: \"\\uC548\\uB155\",\n    cols: \"1\",\n    postId: postId\n  }))))));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (Home);\n\n//# sourceURL=webpack://frontend/./src/Containers/Home/Home.js?");
+eval("__webpack_require__.r(__webpack_exports__);\n/* harmony export */ __webpack_require__.d(__webpack_exports__, {\n/* harmony export */   \"default\": () => (__WEBPACK_DEFAULT_EXPORT__)\n/* harmony export */ });\n/* harmony import */ var react__WEBPACK_IMPORTED_MODULE_0__ = __webpack_require__(/*! react */ \"./node_modules/react/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1__ = __webpack_require__(/*! axios */ \"./node_modules/axios/index.js\");\n/* harmony import */ var axios__WEBPACK_IMPORTED_MODULE_1___default = /*#__PURE__*/__webpack_require__.n(axios__WEBPACK_IMPORTED_MODULE_1__);\n/* harmony import */ var _Home_elements__WEBPACK_IMPORTED_MODULE_2__ = __webpack_require__(/*! ./Home.elements */ \"./src/Containers/Home/Home.elements.js\");\n/* harmony import */ var _globalStyles__WEBPACK_IMPORTED_MODULE_3__ = __webpack_require__(/*! ../../globalStyles */ \"./src/globalStyles.js\");\n/* harmony import */ var _Components__WEBPACK_IMPORTED_MODULE_4__ = __webpack_require__(/*! ../../Components */ \"./src/Components/index.js\");\nfunction _slicedToArray(arr, i) { return _arrayWithHoles(arr) || _iterableToArrayLimit(arr, i) || _unsupportedIterableToArray(arr, i) || _nonIterableRest(); }\n\nfunction _nonIterableRest() { throw new TypeError(\"Invalid attempt to destructure non-iterable instance.\\nIn order to be iterable, non-array objects must have a [Symbol.iterator]() method.\"); }\n\nfunction _unsupportedIterableToArray(o, minLen) { if (!o) return; if (typeof o === \"string\") return _arrayLikeToArray(o, minLen); var n = Object.prototype.toString.call(o).slice(8, -1); if (n === \"Object\" && o.constructor) n = o.constructor.name; if (n === \"Map\" || n === \"Set\") return Array.from(o); if (n === \"Arguments\" || /^(?:Ui|I)nt(?:8|16|32)(?:Clamped)?Array$/.test(n)) return _arrayLikeToArray(o, minLen); }\n\nfunction _arrayLikeToArray(arr, len) { if (len == null || len > arr.length) len = arr.length; for (var i = 0, arr2 = new Array(len); i < len; i++) { arr2[i] = arr[i]; } return arr2; }\n\nfunction _iterableToArrayLimit(arr, i) { var _i = arr == null ? null : typeof Symbol !== \"undefined\" && arr[Symbol.iterator] || arr[\"@@iterator\"]; if (_i == null) return; var _arr = []; var _n = true; var _d = false; var _s, _e; try { for (_i = _i.call(arr); !(_n = (_s = _i.next()).done); _n = true) { _arr.push(_s.value); if (i && _arr.length === i) break; } } catch (err) { _d = true; _e = err; } finally { try { if (!_n && _i[\"return\"] != null) _i[\"return\"](); } finally { if (_d) throw _e; } } return _arr; }\n\nfunction _arrayWithHoles(arr) { if (Array.isArray(arr)) return arr; }\n\n\n\n\n\n\n\nvar Home = function Home(_ref) {\n  var currUser = _ref.currUser,\n      token = _ref.token,\n      userId = _ref.userId;\n\n  var _useState = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(false),\n      _useState2 = _slicedToArray(_useState, 2),\n      loaded = _useState2[0],\n      setLoad = _useState2[1];\n\n  var _useState3 = (0,react__WEBPACK_IMPORTED_MODULE_0__.useState)(\"Loading Content\"),\n      _useState4 = _slicedToArray(_useState3, 2),\n      placeholder = _useState4[0],\n      setPlaceholder = _useState4[1];\n\n  return /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(react__WEBPACK_IMPORTED_MODULE_0__.Fragment, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.Background, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.TodayQuestionSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.TodayQuestion, {\n    currUser: currUser,\n    token: token,\n    userId: userId\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.CreatePageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.CreateCardLeft, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperLeft, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB2F9\\uC2E0\\uC758 \\uC774\\uC57C\\uAE30\\uB97C \\uB4E4\\uB824\\uC8FC\\uC138\\uC694. \\uC5B4\\uB5A4 \\uB0B4\\uC6A9\\uC774\\uB4E0 \\uC88B\\uC544\\uC694. \\uC544\\uC8FC \\uC0AC\\uC18C\\uD55C \\uAC83\\uBD80\\uD130 \\uAE4A\\uC740 \\uC18D\\uB9C8\\uC74C\\uAE4C\\uC9C0, \\uB5A0\\uC624\\uB974\\uB294 \\uB300\\uB85C \\uC801\\uC5B4\\uBCFC\\uAE4C\\uC694?\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.CreateCard, null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.CreateCardRight, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperQuestionArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperRight, null, \"\\uC9C8\\uBB38\\uC744 \\uB4E3\\uACE0,\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB5A0\\uC624\\uB974\\uB294 \\uC74C\\uC545\\uC774\\uB098 \\uC774\\uC57C\\uAE30\\uAC00 \\uC788\\uB098\\uC694?\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB2F9\\uC2E0\\uC758 \\uC774\\uC57C\\uAE30\\uC5D0 \\uC5B4\\uC6B8\\uB9AC\\uB294 \\uC74C\\uC545\\uC744\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uC9C1\\uC811 \\uACE8\\uB77C\\uC8FC\\uC138\\uC694.\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.DoneButtonArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.HelperRight, null, \"\\uC774\\uC81C \\uB2F9\\uC2E0\\uC758 \\uC774\\uC57C\\uAE30\\uB97C\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), \"\\uB2E4\\uB978 \\uC0AC\\uB78C\\uB4E4\\uACFC \\uB098\\uB220\\uBCF4\\uC138\\uC694!\", /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(\"br\", null)), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_globalStyles__WEBPACK_IMPORTED_MODULE_3__.ButtonWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_globalStyles__WEBPACK_IMPORTED_MODULE_3__.DefaultButton, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_globalStyles__WEBPACK_IMPORTED_MODULE_3__.ButtonIconArea, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_globalStyles__WEBPACK_IMPORTED_MODULE_3__.ButtonIcon, {\n    className: \"fa fa-check\"\n  })), \"\\uB2E4 \\uC37C\\uC5B4\\uC694\"))))), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.QuestionListPageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeaderContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeader, null, \"\\uB2E4\\uB978 \\uC9C8\\uBB38 \\uB458\\uB7EC\\uBCF4\\uAE30\"), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.ToQuestionListButton, null, \"\\uBAA8\\uB4E0 \\uC9C8\\uBB38 \\uBCF4\\uAE30 >\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.QuestionCardGrid, {\n    currUser: currUser,\n    token: token,\n    userId: userId\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicListPageSection, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeaderContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.PageHeader, null, \"\\uCD5C\\uADFC\\uC5D0 \\uC120\\uD0DD\\uB41C \\uC74C\\uC545\")), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicListContainer, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicCardWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.MusicCard, {\n    title: \"\\uC548\\uB155\",\n    artist: \"\\uC548\\uB155\",\n    source: \"\\uC548\\uB155\",\n    link: \"\\uC548\\uB155\",\n    cover: \"\\uC548\\uB155\",\n    cols: \"1\",\n    postId: \"1\"\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicCardWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.MusicCard, {\n    title: \"\\uC548\\uB155\",\n    artist: \"\\uC548\\uB155\",\n    source: \"\\uC548\\uB155\",\n    link: \"\\uC548\\uB155\",\n    cover: \"\\uC548\\uB155\",\n    cols: \"1\",\n    postId: \"1\"\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicCardWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.MusicCard, {\n    title: \"\\uC548\\uB155\",\n    artist: \"\\uC548\\uB155\",\n    source: \"\\uC548\\uB155\",\n    link: \"\\uC548\\uB155\",\n    cover: \"\\uC548\\uB155\",\n    cols: \"1\",\n    postId: \"1\"\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicCardWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.MusicCard, {\n    title: \"\\uC548\\uB155\",\n    artist: \"\\uC548\\uB155\",\n    source: \"\\uC548\\uB155\",\n    link: \"\\uC548\\uB155\",\n    cover: \"\\uC548\\uB155\",\n    cols: \"1\",\n    postId: \"1\"\n  })), /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Home_elements__WEBPACK_IMPORTED_MODULE_2__.MusicCardWrapper, null, /*#__PURE__*/react__WEBPACK_IMPORTED_MODULE_0__.createElement(_Components__WEBPACK_IMPORTED_MODULE_4__.MusicCard, {\n    title: \"\\uC548\\uB155\",\n    artist: \"\\uC548\\uB155\",\n    source: \"\\uC548\\uB155\",\n    link: \"\\uC548\\uB155\",\n    cover: \"\\uC548\\uB155\",\n    cols: \"1\",\n    postId: \"1\"\n  }))))));\n};\n\n/* harmony default export */ const __WEBPACK_DEFAULT_EXPORT__ = (Home);\n\n//# sourceURL=webpack://frontend/./src/Containers/Home/Home.js?");
 
 /***/ }),
